pyproject.toml
[project]
name = "librarifastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Vadim Bakhtiarov"}
]
packages = [
    { include = "lib_api" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "greenlet (>=3.2.2,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fastapi-pagination (>=0.13.1,<0.14.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi-csrf-protect (>=1.0.3,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
loguru = "^0.7.3"
isort = "^6.0.1"
ruff = "^0.11.10"
